Force integer value.

+++ log.php
@@ -83,7 +83,7 @@ if (isset($_POST['stats']) || isset($_GET['stats'])) {
             } elseif ($stats[0] == 'error') {
                 $errors[] = $stats[1];
             }
-            $microseconds = 60000000/$log_rate-round((microtime(true)-$get_stats_start)*1000000);
+            $microseconds = (int) 60000000/$log_rate-round((microtime(true)-$get_stats_start)*1000000);
             if ($microseconds > 0) {
                 usleep($microseconds);
             }
